... reconciling a bunch of different API changes in the process. Aaaargh!

+++ src/os2datascanner/engine2/model/utilities.py
@@ -1,6 +1,7 @@
 from os import rmdir, remove
 import os.path
 from tempfile import mkdtemp
+from contextlib import suppress
 
 class NamedTemporaryResource:
     def __init__(self, name):
@@ -8,14 +9,20 @@ class NamedTemporaryResource:
         self._dir = None
 
     def open(self, mode):
-        if self._dir is None:
-            self._dir = mkdtemp()
         return open(self.get_path(), mode)
 
     def get_path(self) -> str:
+        if self._dir is None:
+            self._dir = mkdtemp()
         return os.path.join(self._dir, self._name)
 
-    def finished(self):
-        remove(self.get_path())
-        rmdir(self._dir)
-        self._dir = None
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, backtrace):
+        if self._dir:
+            with suppress(FileNotFoundError):
+                remove(self.get_path())
+            with suppress(FileNotFoundError):
+                rmdir(self._dir)
+            self._dir = None
